Clean it up quite a lot
