+++ src/testing/test-utils.ts
@@ -4,13 +4,7 @@ import { buildStores } from 'app/inventory/store/d2-store-factory';
 import { downloadManifestComponents } from 'app/manifest/manifest-service-json';
 import { humanBytes } from 'app/storage/human-bytes';
 import { delay } from 'app/utils/promises';
-import {
-  AllDestinyManifestComponents,
-  DestinyManifest,
-  DestinyProfileResponse,
-  DestinyVendorsResponse,
-  ServerResponse,
-} from 'bungie-api-ts/destiny2';
+import { AllDestinyManifestComponents, DestinyManifest } from 'bungie-api-ts/destiny2';
 import { F_OK } from 'constants';
 import { maxBy, once } from 'es-toolkit';
 import i18next from 'i18next';
@@ -31,8 +25,7 @@ import zhCHT from 'locale/zhCHT.json';
 import fs from 'node:fs/promises';
 import path from 'node:path';
 import { getManifest as d2GetManifest } from '../app/bungie-api/destiny2-api';
-import profile from './data/profile-2024-06-13.json';
-import vendors from './data/vendors-2024-06-13.json';
+import { getTestProfile } from './test-profile';
 
 /**
  * Get the current manifest as JSON. Downloads the manifest if not cached.
@@ -132,11 +125,6 @@ export const testAccount = {
   platforms: [1, 3, 5, 2],
   lastPlayed: '2021-05-08T03:34:26.000Z',
 };
-
-export const getTestProfile = () =>
-  (profile as unknown as ServerResponse<DestinyProfileResponse>).Response;
-export const getTestVendors = () =>
-  (vendors as unknown as ServerResponse<DestinyVendorsResponse>).Response;
 
 export const getTestStores = once(async () => {
   const manifest = await getTestDefinitions();
